requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
yfinance==0.2.35
anthropic==0.17.0
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
pandas==2.2.0
pydantic-settings==2.1.0

